feat(tauri): 任务栏编排与数据层打磨（Windows 适配后续）
任务栏（taskband.rs）：
- 上行文字色跟随系统任务栏（ColorStyle::Default，深浅色自适应）；
  分组自定义色/全局上行色仍优先（top_color_style 解析链）
- 实例右键菜单新增「隐藏「分组名」」项（hide_group_from_menu → 写
  menubarHiddenGroups，语义与设置页取消勾选一致；总览不可隐藏）
- 样式下发精简：固定上下两行，不再读 menubar_layout/equal_font_size

插件依赖：multiline-taskband pin rev 9597d41 → tag v1.0.0
（首个 stable tag，仅 Windows 渲染修复，Rust/JS API 无变化），
适配说明文档同步改版（能力差距评估结论：全部不处理）。

数据层增强：
- fundmnfinfo：自算估值失败时兜底 fund123 实时估值
  （calc_fund_estimate_change + fund123_estimate_fallback）；
  板块推断按需刷新 refresh_sectors_if_needed
- fund123：盘中估值分时 get_fund_estimate_intraday（详情弹窗用）
- history/fundname：resolve_fund 与按名称选基金（pick_fund_by_name）重构
- market：指数历史 K 线兜底链整理（腾讯/新浪 CN/新浪 US/东财/新浪外盘）
- portfolio：normalize_fund_map 归一化；theme：板块词库扩展
- http/providers/mod：公共请求封装（eastmoney_fund_get 等）

设置页（OptionsApp）：任务栏分区文案更新（系统色说明/停靠位置）；
fundOps 补 menubarGroupSides / menubarEdgeMargins 归一化写入。

+++ apps/tauri/src-tauri/src/window.rs
@@ -306,9 +306,9 @@ fn position_below(app: &AppHandle, win: &WebviewWindow, rect: (f64, f64, f64, f6
             let p = m.position();
             let s = m.size();
             cx >= p.x as f64
-                && cx < (p.x + s.width) as f64
+                && cx < (p.x + s.width as i32) as f64
                 && cy >= p.y as f64
-                && cy < (p.y + s.height) as f64
+                && cy < (p.y + s.height as i32) as f64
         })
         .or_else(|| app.primary_monitor().ok().flatten());
     let (mx, my, mw, mh) = monitor
